refctor :: 셀 수정 컴포넌트화 시키고 임포트 및 푸터 임포트
